third_party/blink/web_tests/fast/filesystem/workers/null-arguments-expected.txt

[Worker] This test tries calling various filesystem functions with null arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/null-arguments-worker.js
PASS [Worker] fileSystem.root.moveTo(null, 'x') threw exception TypeError: Failed to execute 'moveTo' on 'EntrySync': parameter 1 is not of type 'DirectoryEntrySync'..
PASS [Worker] fileSystem.root.copyTo(null, 'x') threw exception TypeError: Failed to execute 'copyTo' on 'EntrySync': parameter 1 is not of type 'DirectoryEntrySync'..
PASS [Worker] writer.write(null) threw exception TypeError: Failed to execute 'write' on 'FileWriterSync': parameter 1 is not of type 'Blob'..
PASS successfullyParsed is true

TEST COMPLETE
